Moved logic for tracking to the page route and updated to track what we want (vs duplicate requests)
