+++ src/pages/outcome/[rowKey].tsx
@@ -1,9 +1,12 @@
-import { useGameInfoContext } from 'components/Contexts';
+import { useEngageTracker, useGameInfoContext } from 'components/Contexts';
 import { OutcomePanel, OutcomeUnavailable } from 'components/Outcomes';
 import { Layout, SingleColumnLayout } from 'components/ui';
+import * as GTag from 'lib/GTag';
 import { MediaService } from 'lib/MediaService';
 import { IAnswer, IImage } from 'models';
+import { OutcomeConditions, TargetProduct } from 'models/OutcomeConditions';
 import { GetStaticProps } from 'next';
+import { useRouter } from 'next/router';
 import { FC, useEffect } from 'react';
 import { AzureTableService } from 'services/AzureTable/AzureTableService';
 
@@ -61,25 +64,60 @@ export const getStaticProps: GetStaticProps = async (context) => {
 };
 
 const OutcomeHashPage: FC<OutcomeHashPageProps> = (props) => {
-  const gameInfo = useGameInfoContext();
+  const router = useRouter();
+  const gameInfoContext = useGameInfoContext();
+  const tracker = useEngageTracker();
 
   useEffect(() => {
+    // This will track all outcome pages
+    // props can be empty on first render, but updates later, this prevents two tracking events from firing
+    if (props.error || props.answers) {
+      tracker.TrackPageView(
+        { page: router.asPath, channel: 'WEB', language: 'EN', currency: 'USD' },
+        {
+          answers: JSON.stringify(props.answers),
+          error: props.error,
+        }
+      );
+    }
+
     if (props.avatar) {
-      gameInfo.updateAvatar(props.avatar);
+      gameInfoContext.updateAvatar(props.avatar);
     }
 
     if (props.theme) {
-      gameInfo.updateTheme(props.theme);
+      gameInfoContext.updateTheme(props.theme);
     }
 
     if (props.persona) {
-      gameInfo.updatePersona(props.persona);
+      gameInfoContext.updatePersona(props.persona);
     }
 
     if (props.answers) {
-      gameInfo.updateAnswers(props.answers);
+      if (!gameInfoContext.answers || gameInfoContext.answers.length === 0) {
+        // If answers already exist, don't update them (Means this is someone completing the quiz vs a page view)
+        gameInfoContext.updateAnswers(props.answers);
+      }
+
+      GTag.event('outcome_answers', 'Answers', JSON.stringify(props.answers));
     }
   }, [props]);
+
+  useEffect(() => {
+    // Since Answers is async in nature because they will get updated from the useEffect based on props, we will wait to log event until gameInfoContext.answers is updated
+    // Could refactor outcomeconditions though in the future, since it doesn't need gameInfoContext, it just needs answers, which you could pass to it.
+    if (gameInfoContext.answers && gameInfoContext.answers.length > 0) {
+      let outcomeConditions = new OutcomeConditions(gameInfoContext);
+      const requiredProducts: TargetProduct[] = outcomeConditions.requiredProducts();
+
+      if (requiredProducts) {
+        requiredProducts.forEach((product) => {
+          tracker.TrackEvent('outcome_required_product', { requiredProduct: product });
+          GTag.event('outcome_required_product', product, product);
+        });
+      }
+    }
+  }, [gameInfoContext.answers]);
 
   if (props.error) {
     return (
@@ -90,7 +128,7 @@ const OutcomeHashPage: FC<OutcomeHashPageProps> = (props) => {
           showResetButton={true}
           showFeedbackButton={true}
           backgroundImage={
-            gameInfo.theme?.chakraTheme == 'fantasy'
+            gameInfoContext.theme?.chakraTheme == 'fantasy'
               ? 'https://mms-delivery.sitecorecloud.io/api/media/v2/delivery/df4c80ea-db67-49f8-bcd3-08daadeee4f5/182bc6d196aa465cbf9b614ff2883eb4'
               : 'https://mms-delivery.sitecorecloud.io/api/media/v2/delivery/df4c80ea-db67-49f8-bcd3-08daadeee4f5/1821f8838e284d6fad1d483d41877aba'
           }
